+++ types/ramda/test/merge-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  R.merge({ name: 'fred', age: 10 }, { age: 40 });
+  // => { 'name': 'fred', 'age': 40 }
+};
