+++ types/ramda/test/mergeDeepRight-tests.ts
@@ -1,0 +1,5 @@
+import * as R from 'ramda';
+
+() => {
+  const a = R.mergeDeepRight({ foo: { bar: 1 } }, { foo: { bar: 2 } }); // => {foo: bar: 2}}
+};
