+++ types/ramda/test/modulo-tests.ts
@@ -1,0 +1,8 @@
+import * as R from 'ramda';
+
+() => {
+  R.modulo(17, 3); // => 2
+  // JS behavior:
+  R.modulo(-17, 3); // => -2
+  R.modulo(17, -3); // => 2
+};
