+++ types/ramda/test/negate-tests.ts
@@ -1,0 +1,5 @@
+import * as R from 'ramda';
+
+() => {
+  R.negate(42); // => -42
+};
