+++ types/ramda/tsconfig.json
@@ -132,6 +132,26 @@
         "test/mathMod-tests.ts",
         "test/max-tests.ts",
         "test/maxBy-tests.ts",
-        "test/mean-tests.ts"
+        "test/mean-tests.ts",
+        "test/median-tests.ts",
+        "test/memoizeWith-tests.ts",
+        "test/merge-tests.ts",
+        "test/mergeAll-tests.ts",
+        "test/mergeDeepLeft-tests.ts",
+        "test/mergeDeepRight-tests.ts",
+        "test/mergeDeepWith-tests.ts",
+        "test/mergeDeepWithKey-tests.ts",
+        "test/mergeLeft-tests.ts",
+        "test/mergeRight-tests.ts",
+        "test/mergeWith-tests.ts",
+        "test/mergeWithKey-tests.ts",
+        "test/min-tests.ts",
+        "test/minBy-tests.ts",
+        "test/modulo-tests.ts",
+        "test/move-tests.ts",
+        "test/multiply-tests.ts",
+        "test/nAry-tests.ts",
+        "test/negate-tests.ts",
+        "test/none-tests.ts"
     ]
 }
